// src/app/modules/billing/random/models/random-waybills-result.model.ts
export interface RandomWaybillsResult {
    baseAmount: number;
    cgstAmount: number;
    createdDate: Date;
    deliveryDate: Date;
    destinationCode: string;
    destinationId: number;
    igstAmount: number;
    originCode: string;
    originId: number;
    outstandingAmount: number;
    pickupDate: Date;
    pkg: number;
    sfxCode: string;
    sgstAmount: number;
    ttlTaxAmount: number;
    waybillId: number;
    waybillNumber: string;
    weight: number;

}